pyproject.toml
[project]
name = "agrivision-pro-v3"
version = "3.0.0"
description = "AgriVision Pro - Satellite Vegetation Analysis Platform"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "AgriVision Pro Team"}
]

dependencies = [
    "streamlit>=1.28.0",
    "earthengine-api>=0.1.370",
    "geemap>=0.30.0",
    "folium>=0.14.0",
    "streamlit-folium>=0.15.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "plotly>=5.18.0",
    "scipy>=1.11.0",
    "Pillow>=10.0.0",
    "google-auth>=2.23.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
